+++ src/popover.ts
@@ -1,5 +1,4 @@
 import { observePopoversMutations, popovers } from './observer.js';
-import { queryAncestorAll } from './triggers.js';
 
 export function isSupported() {
   return (
@@ -8,6 +7,39 @@ export function isSupported() {
     'popover' in HTMLElement.prototype
   );
 }
+
+const closestAncestor: (selector: string, target: Element) => Element | null = (
+  selector: string,
+  target: Element,
+) => {
+  const found = target.closest(selector);
+
+  if (found) {
+    return found;
+  }
+
+  const root = target.getRootNode();
+
+  if (root === document || !(root instanceof ShadowRoot)) {
+    return null;
+  }
+
+  return closestAncestor(selector, root.host);
+};
+
+const queryAncestorAll = (
+  element: Element,
+  selector: string,
+  popovers: Element[] = [],
+): Element[] => {
+  // there could be multiple popovers nested inside each other
+  const ancestor = closestAncestor(selector, element);
+  const parent =
+    ancestor?.parentElement || (ancestor?.getRootNode() as ShadowRoot)?.host;
+  return ancestor && parent
+    ? queryAncestorAll(parent, selector, [ancestor, ...popovers])
+    : popovers;
+};
 
 function patchAttachShadow(callback: (shadowRoot: ShadowRoot) => void) {
   const originalAttachShadow = Element.prototype.attachShadow;
